refactor: package 재구성
domain
-> domain.aggregate
-> domain.code
-> domain.entity
-> domain.repository
-> domain.vo

Resolve:
See Also:
