+++ src/main/java/qna/domain/aggregate/AnswerGroup.java
@@ -1,6 +1,5 @@
-package qna.domain;
+package qna.domain.aggregate;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -8,6 +7,9 @@ import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.FetchType;
 import javax.persistence.OneToMany;
+import qna.domain.entity.Answer;
+import qna.domain.entity.DeleteHistory;
+import qna.domain.entity.User;
 
 @Embeddable
 public class AnswerGroup {
