+++ src/main/java/qna/domain/entity/Question.java
@@ -1,6 +1,5 @@
-package qna.domain;
+package qna.domain.entity;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
 import javax.persistence.Column;
@@ -12,7 +11,9 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import qna.CannotDeleteException;
+import qna.exception.CannotDeleteException;
+import qna.domain.aggregate.AnswerGroup;
+import qna.domain.aggregate.DeleteHistoryGroup;
 import qna.domain.vo.Contents;
 import qna.domain.vo.Title;
