+++ src/main/java/qna/domain/repository/AnswerRepository.java
@@ -1,7 +1,8 @@
-package qna.domain;
+package qna.domain.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.entity.Answer;
 
 
 public interface AnswerRepository extends JpaRepository<Answer, Long> {
