+++ src/main/java/qna/domain/repository/DeleteHistoryRepository.java
@@ -1,6 +1,7 @@
-package qna.domain;
+package qna.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.entity.DeleteHistory;
 
 public interface DeleteHistoryRepository extends JpaRepository<DeleteHistory, Long> {
