+++ src/main/java/qna/domain/repository/QuestionRepository.java
@@ -1,7 +1,8 @@
-package qna.domain;
+package qna.domain.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.entity.Question;
 
 public interface QuestionRepository extends JpaRepository<Question, Long> {
