+++ src/main/java/qna/domain/repository/UserRepository.java
@@ -1,7 +1,8 @@
-package qna.domain;
+package qna.domain.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.entity.User;
 import qna.domain.vo.UserId;
 
 public interface UserRepository extends JpaRepository<User, Long> {
