+++ src/main/java/qna/domain/vo/Title.java
@@ -15,7 +15,7 @@ public class Title {
 	protected Title() {
 	}
 
-	protected Title(String title) {
+	private Title(String title) {
 		this.title = title;
 	}
