+++ src/main/java/qna/service/DeleteHistoryService.java
@@ -4,8 +4,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import qna.domain.DeleteHistory;
-import qna.domain.DeleteHistoryRepository;
+import qna.domain.entity.DeleteHistory;
+import qna.domain.repository.DeleteHistoryRepository;
 
 @Service
 public class DeleteHistoryService {
