+++ src/test/java/qna/domain/aggregate/AnswerGroupTest.java
@@ -1,10 +1,11 @@
-package qna.domain;
+package qna.domain.aggregate;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import java.util.ArrayList;
 import java.util.Objects;
 import org.junit.jupiter.api.Test;
+import qna.domain.aggregate.AnswerGroup;
 
 public class AnswerGroupTest {
