+++ src/test/java/qna/domain/aggregate/DeleteHistoryGroupTest.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.aggregate;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -7,6 +7,10 @@ import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import qna.domain.entity.Answer;
+import qna.domain.entity.DeleteHistory;
+import qna.domain.entity.Question;
+import qna.domain.entity.User;
 
 public class DeleteHistoryGroupTest {
