+++ src/test/java/qna/domain/entity/AnswerTest.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
