+++ src/test/java/qna/domain/entity/QuestionTest.java
@@ -1,11 +1,10 @@
-package qna.domain;
+package qna.domain.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import qna.CannotDeleteException;
 
 public class QuestionTest {
