+++ src/test/java/qna/domain/repository/DeleteHistoryRepositoryTest.java
@@ -1,14 +1,17 @@
-package qna.domain;
+package qna.domain.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import java.time.LocalDateTime;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import qna.domain.entity.Answer;
+import qna.domain.entity.DeleteHistory;
+import qna.domain.entity.Question;
+import qna.domain.entity.User;
 
 @DataJpaTest
 public class DeleteHistoryRepositoryTest {
