+++ src/test/java/qna/domain/repository/UserRepositoryTest.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import qna.domain.entity.User;
 
 @DataJpaTest
 public class UserRepositoryTest {
